HDFS-17874. (Folloup) Fix package issue by moving SampleStep to a hadoop package
