+++ hadoop-hdfs-project/hadoop-hdfs/src/test/java/org/apache/hadoop/test/SampleStep.java
@@ -14,7 +14,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package sample;
+package org.apache.hadoop.test;
 
 import org.apache.hadoop.hdfs.server.diskbalancer.datamodel.DiskBalancerVolume;
 import org.apache.hadoop.hdfs.server.diskbalancer.planner.Step;
